move default taskManager to file reader for parity
